web/requirements.txt
asyncpg==0.30.0
fastapi==0.115.6
sqladmin==0.20.1
sqlalchemy==2.0.37
uvicorn==0.34.0
pika==1.3.2
python-dotenv==1.0.1
pydantic==2.10.5
starlette==0.41.3
greenlet==3.1.1
ruff==0.9.10
aiosmtplib==4.0.0
python-jose==3.4.0
itsdangerous==2.2.0
httpx==0.28.1
pytest==8.3.5